Реализовал базу данных в виде мапы. Удалил лишний абстрактный класс и обновил список задач на выполнение.
