+++ README.md
@@ -1,6 +1,9 @@
 # Что нужно реализовать в этом проекте?
-1) Реализовать работу со структурой данных типа Map
-2) Реализовать запись пар слов в файл
+1) Реализовать работу со структурой данных типа Map +
+2) Реализовать запись пар слов в файл +
 3) Реализовать меню пользователя +
-4) Организовать функции для проверки пользователя на знание слов
-5) Организовать базу данных записанных слов
+4) Организовать функции для проверки пользователя на знание слов +
+5) Организовать базу данных записанных слов +
+6) Считывать из файла уже записанные слова в Мапу
+7) Подсчитывать количество правильно отгаданных слов и выдавать на обозрение
+8) Удалять правильно отгаданные слова (допустим после трех успешных попыток отгадать слово)
