--- src/main/java/Word.java
@@ -1,20 +1,0 @@
-public abstract class Word {
-    private String english;
-    private String russian;
-
-    public String getEnglish() {
-        return english;
-    }
-
-    public void setEnglish(String english) {
-        this.english = english;
-    }
-
-    public String getRussian() {
-        return russian;
-    }
-
-    public void setRussian(String russian) {
-        this.russian = russian;
-    }
-}
